ivar_write_analyzer
