+++ lib/rbs_infer/analyzer.rb
@@ -1266,5 +1266,6 @@ require_relative "inference/param_type_inferrer"
 require_relative "project/source_index"
 require_relative "project/mixin_index"
 require_relative "signatures/steep_bridge"
+require_relative "signatures/steep_bridge/ivar_write_analyzer"
 require_relative "project/source_expanders"
 require_relative "project/self_type_annotators"
